Split heif_color_profile() to separate module
Simplify conditional compilation.
