+++ src/image.cc
@@ -29,6 +29,7 @@
 
 #include "collect-table.h"
 #include "collect.h"
+#include "color-man-heif.h"
 #include "color-man.h"
 #include "compat.h"
 #include "exif.h"
