+++ src/meson.build
@@ -47,6 +47,8 @@ main_sources = files('advanced-exif.cc',
 'collect-table.h',
 'color-man.cc',
 'color-man.h',
+'color-man-heif.cc',
+'color-man-heif.h',
 'command-line-handling.cc',
 'command-line-handling.h',
 'compat.cc',
